refactored the back package directory, added java doc to the classes responsible for the "Back" Use Case
