+++ src/app/DisplayNotesUseCaseFactory.java
@@ -1,12 +1,11 @@
 package app;
 
-import data_access.NoteDataAccessObject;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.note.DeleteNoteController;
 import interface_adapter.note.DeleteNotePresenter;
 import interface_adapter.note.NoteViewModel;
-import interface_adapter.note.back.BackController;
-import interface_adapter.note.back.BackPresenter;
+import interface_adapter.back.BackController;
+import interface_adapter.back.BackPresenter;
 import interface_adapter.note.display_notes.DisplayNotesController;
 import interface_adapter.note.display_notes.DisplayNotesPresenter;
 import interface_adapter.note.display_notes.DisplayNotesViewModel;
