+++ src/app/DisplayQuizzesUseCaseFactory.java
@@ -4,8 +4,8 @@ import interface_adapter.ViewManagerModel;
 
 
 import interface_adapter.note.NoteViewModel;
-import interface_adapter.note.back.BackController;
-import interface_adapter.note.back.BackPresenter;
+import interface_adapter.back.BackController;
+import interface_adapter.back.BackPresenter;
 import interface_adapter.question.QuestionViewModel;
 import interface_adapter.quiz.delete.DeleteQuizController;
 import interface_adapter.quiz.delete.DeleteQuizPresenter;
@@ -13,7 +13,6 @@ import interface_adapter.quiz.display.DisplayQuizzesController;
 import interface_adapter.quiz.display.DisplayQuizzesPresenter;
 import interface_adapter.quiz.display.DisplayQuizzesViewModel;
 import interface_adapter.quiz.take_quiz.TakeQuizController;
-import interface_adapter.quiz.take_quiz.TakeQuizPresenter;
 import interface_adapter.quiz.take_quiz.TakeQuizViewModel;
 import use_case.quiz.delete.DeleteQuizInputBoundary;
 import use_case.quiz.delete.DeleteQuizInteractor;
@@ -22,10 +21,6 @@ import use_case.quiz.display.DisplayQuizzesInputBoundary;
 import use_case.quiz.display.DisplayQuizzesInteractor;
 import use_case.quiz.display.DisplayQuizzesOutputBoundary;
 import use_case.quiz.QuizDataAccessInterface;
-import use_case.quiz.take_quiz.TakeQuizInputBoundary;
-import use_case.quiz.take_quiz.TakeQuizInputData;
-import use_case.quiz.take_quiz.TakeQuizInteractor;
-import use_case.quiz.take_quiz.TakeQuizOutputBoundary;
 import view.DisplayQuizzesView;
 
 public class DisplayQuizzesUseCaseFactory {
