+++ src/view/DisplayQuizzesView.java
@@ -1,21 +1,13 @@
 package view;
 
-import data_access.QuizDataAccessObject;
-import interface_adapter.ViewManagerModel;
-import interface_adapter.note.back.BackController;
+import interface_adapter.back.BackController;
 import interface_adapter.quiz.delete.DeleteQuizController;
 import interface_adapter.quiz.display.DisplayQuizzesController;
 import interface_adapter.quiz.display.DisplayQuizzesState;
 import interface_adapter.quiz.display.DisplayQuizzesViewModel;
 import interface_adapter.quiz.take_quiz.TakeQuizController;
-import interface_adapter.quiz.take_quiz.TakeQuizPresenter;
-import interface_adapter.quiz.take_quiz.TakeQuizState;
 import interface_adapter.quiz.take_quiz.TakeQuizViewModel;
 import org.bson.types.ObjectId;
-import use_case.quiz.QuizDataAccessInterface;
-import use_case.quiz.take_quiz.TakeQuizInputBoundary;
-import use_case.quiz.take_quiz.TakeQuizInteractor;
-import use_case.quiz.take_quiz.TakeQuizOutputBoundary;
 
 import javax.swing.*;
 import javax.swing.event.ListSelectionEvent;
